feat(chat): warm daemon pull for task and project chat
